+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Package-specific changes not released in any SDK will be added here just before 
 
 ### 📚 3rd party library updates
 
+- Updated `react-native-webview` from `11.13.0` to `11.15.0`.
+
 ### 🛠 Breaking changes
 
 ### 🎉 New features